Update Program.cs
Perhaps this is simpler? Instead of passing a large number of command-line arguments or reading a job file with details of notebooks to collect, this program executes once per notebook. Create a batch file to collect multiple notebooks. If one item in the batch fails, it should not disturb the rest of the batch.

Alternatively, look into using something specifically designed for the purpose, like wget or cURL.